+++ dst/app/controllers/api/performances.js
@@ -1,17 +1,19 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.getTicketInfo = exports.getTargetReservationsForRefund = exports.updateOnlineStatus = exports.getUnitPriceByAcceptedOffer = void 0;
 /**
  * パフォーマンスAPIコントローラー
  */
-const cinerinoapi = require("@cinerino/api-nodejs-client");
+const cinerinoapi = require("@cinerino/sdk");
 const tttsapi = require("@motionpicture/ttts-api-nodejs-client");
 const conf = require("config");
 const createDebug = require("debug");
@@ -341,7 +343,7 @@ function createEmail(req, res, order, notice) {
         // その場で送信ではなく、DBにタスクを登録
         const taskAttributes = {
             name: tttsapi.factory.taskName.SendEmailMessage,
-            project: order.project,
+            project: { typeOf: order.project.typeOf, id: order.project.id },
             status: tttsapi.factory.taskStatus.Ready,
             runsAt: new Date(),
             remainingNumberOfTries: 10,