+++ dst/app/controllers/language.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.update = void 0;
 const _ = require("underscore");
 /**
  * 言語切り替え