+++ src/app/controllers/api/performances/suspended.ts
@@ -1,7 +1,7 @@
 /**
  * 販売停止パフォーマンスAPIコントローラー
  */
-import * as cinerinoapi from '@cinerino/api-nodejs-client';
+import * as cinerinoapi from '@cinerino/sdk';
 import * as tttsapi from '@motionpicture/ttts-api-nodejs-client';
 import * as createDebug from 'debug';
 import * as Email from 'email-templates';