+++ src/app/controllers/staff/mypage.ts
@@ -1,7 +1,7 @@
 /**
  * マイページコントローラー
  */
-import * as cinerinoapi from '@cinerino/api-nodejs-client';
+import * as cinerinoapi from '@cinerino/sdk';
 // import * as tttsapi from '@motionpicture/ttts-api-nodejs-client';
 
 import * as createDebug from 'debug';