+++ src/index.ts
@@ -71,12 +71,10 @@ function onError(error: any) {
             // tslint:disable-next-line:no-console
             console.error(`${bind} requires elevated privileges`);
             process.exit(1);
-            break;
         case 'EADDRINUSE':
             // tslint:disable-next-line:no-console
             console.error(`${bind} is already in use`);
             process.exit(1);
-            break;
         default:
             throw error;
     }